update gps code

+++ Project1/3년차/gps/gps_with_gyro2.cpp
@@ -332,11 +332,11 @@ float get_atmospheric(char *chrBuf, int num)
 double get_gpsData(char *chrBuf, int num)
 {
     double lon; double lat;
-    signed int tmp[8];// int로 변경
+    unsigned short tmp[8];// int로 변경
     unsigned char i;
 
     for(i=0;i<8;i++){
-        tmp[i] = (signed int)chrBuf[i+2];// int로 변경
+        tmp[i] = (unsigned short)chrBuf[i+2];// int로 변경
     }
 
     if (num==1){
